fix extract_db_type_from_uri
